feat: Add `noexcept` specifier to `Acts::Any` (#3743)
In the ESC24 school I learned, that `noexcept` on special member functions, especially on move assignment and move construction can be very impactful on performance with `std::vector`. 

The reasons seems to be the strong exception garantuees of `std::vector`: If the move construction is not `noexcept` it falls back to copy, because it is required to be able to restore the initial state in case of failure of an operation. See for this e.g. the quickbench result on a dummy `String` implementation: https://quick-bench.com/q/5CG81CYyOUm7xSa9ClbyLiPj7SY

I won't have any effect on source links I think because of SBO, but in general would be useful to have here I think.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
	- Enhanced exception safety for the `AnyBase` class.
	- Improved debugging capabilities with verbose output for allocation and deallocation operations.

- **Bug Fixes**
	- Updated method signatures to ensure consistent exception handling based on allocation tracking settings.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->